Se agrega que se pueda cambiar de mes en los registros de gastos

+++ src/app/app/[user]/resume/components/TableExpenses/TableExpenses.tsx
@@ -3,7 +3,7 @@ import { FormatDate } from "@/app/utils/FormatDate"
 import { toast } from "sonner"
 import { ExpenseModel } from "../../hooks/useExpenses"
 
-export default function TableExpenses ({data, refresh}: {data: ExpenseModel[], refresh: any}) {
+export default function TableExpenses ({data, refresh, monthCurrent}: {data: ExpenseModel[], refresh: any, monthCurrent: number}) {
 
   const handleDelete = async (id: number) => {
     const res = await fetch(`/api/expenses?id=${id}`, {
@@ -25,23 +25,29 @@ export default function TableExpenses ({data, refresh}: {data: ExpenseModel[], r
     <div className='max-h-[400px] overflow-y-auto'>
       {
         data.length >= 1 ?
-        data.map((item) => (
-          <div key={item.id} className="py-5 grid grid-cols-3 md:grid-cols-4 items-center border-b text-sm md:text-md hover:bg-[#201D1D] cursor-pointer">
-          <div className="flex gap-2 items-center text-white">
-            <div className="hidden md:block">
-              {/* <FoodIcon /> */}
-              <DeleteConfirmation deleteItem={() => handleDelete(item?.id)} message="¿Deseas eliminar este gasto?" />
-            </div>
-            <div>
-              <span>{item?.categoryname}</span>
-              <p className="block md:hidden opacity-70 text-sm">{item?.description}</p>
-            </div>
-          </div>
-          <p className="text-white opacity-70 hidden md:block">{item?.description !== '' ? item.description : 'Sin descripción'}</p>
-          <span className='text-white opacity-70'>{FormatDate(item?.date)}</span>
-          <span className='text-palette font-medium'>$ {item?.value?.toLocaleString()}</span>
-        </div>
-        )) : <></>
+            data.map((item) => {
+              const dateExpense = new Date(item.date)
+              const monthExpense = dateExpense.getMonth()
+              if (monthCurrent === monthExpense) {
+                return (
+                  <div key={item.id} className="py-5 grid grid-cols-3 md:grid-cols-4 items-center border-b text-sm md:text-md hover:bg-[#201D1D] cursor-pointer">
+                    <div className="flex gap-2 items-center text-white">
+                      <div className="hidden md:block">
+                        {/* <FoodIcon /> */}
+                        <DeleteConfirmation deleteItem={() => handleDelete(item?.id)} message="¿Deseas eliminar este gasto?" />
+                      </div>
+                      <div>
+                        <span>{item?.categoryname}</span>
+                        <p className="block md:hidden opacity-70 text-sm">{item?.description}</p>
+                      </div>
+                    </div>
+                    <p className="text-white opacity-70 hidden md:block">{item?.description !== '' ? item.description : 'Sin descripción'}</p>
+                    <span className='text-white opacity-70'>{FormatDate(item?.date)}</span>
+                    <span className='text-palette font-medium'>$ {item?.value?.toLocaleString()}</span>
+                  </div>
+                )
+              }
+            }) : <></>
       }
       </div>
     </>
